Implemented Split Horizon.
Routers will now not send rtes to a router if the next hop for that rte is is the same
router. This helps a basic class of Routing loops, speeding up convergence time if a router
is deleted. Tested with routers1, 2 and 3, with router 3 disabled once the routers had converged.