Implemented Split Horizon with Poison Reverse
Ran into some issues when assigning the entry to a variable like so:
	poisoned_entry = entry
Updating `poisoned_entry`s cost seemed to change the original entries
value, causing the routing table for that router to change. This was
avoided by creating a new RIPRouteEntry, identical to `entry` apart from
having a metric of infinity.